Worked on adding locations insurance/specialties
There was an error in user_profile_class.js due to a missing ')' on an if statement.

I worked in the database and the PHP files pertaining to the add_insurance.php. I am able to successfully insert multiple insurance and specialties into the database, albeit the stored procedure broke and needs to be fixed (sorry @dmeza38 ).

I also added a new table that has a list of insurances that the user and vendor can choose from.